requirements.txt
requests
uvicorn
fastapi
PyYAML
openpyxl
asyncpg
ormar
psycopg2-binary
alembic
apscheduler
fastapi-utils
mypy
pydantic
SQLAlchemy
pytz
starlette
pip
setuptools
wheel
aiofiles
python-multipart
pandas
loguru
databases
aiosqlite